fix(gen,sim): do modulo in u64 before narrowing to usize
On wasm32 usize is 32-bit, so '(h as usize) % len' truncates the hash
BEFORE the modulo and a non-power-of-two modulo diverges from native
64-bit results - building style picks (and thus door tiles and every
downstream home coordinate) differed between native and wasm builds of
the same world seed. Found by the voxel-village M0-S2 bridge tripwire:
native/wasm world checksums diverged at tick 0 with identical entity
ids. Same class fixed in aspiration's master-trade pick. Native output
is bit-identical (usize is already 64-bit there); wasm now matches
native checksums at tick 0 and tick 1000.

+++ src/sim/aspiration.rs
@@ -124,7 +124,8 @@ pub fn tick_settlement_aspirations(s: &mut Settlement, seed: u64, tick: u64) -> 
                     .filter(|t| SKILLED_TRADES.contains(&t.as_str()))
                     .collect();
                 if !masters.is_empty() {
-                    let t = masters[(rng as usize) % masters.len()].clone();
+                    // Modulo in u64 before narrowing: usize is 32-bit on wasm32.
+                    let t = masters[(rng % masters.len() as u64) as usize].clone();
                     Some(Aspiration::LearnTrade {
                         trade: t,
                         progress: 0.0,
